requirements.txt
argon2-cffi==20.1.0
asn1crypto==1.2.0
attrs==20.1.0
backcall==0.2.0
beautifulsoup4==4.8.1
bleach==3.1.5
blis==0.4.1
boto==2.49.0
boto3==1.10.20
botocore==1.13.20
bz2file==0.98
catalogue==0.0.8
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
elasticsearch==7.9.1
en-core-web-sm==2.2.5
entrypoints==0.3
futures==3.1.1
gensim==3.8.0
idna==2.8
importlib-metadata==0.23
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.14.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server==1.2.0
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.4.5
notebook==6.1.3
numpy==1.17.3
olefile==0.46
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pudb==2019.2
py-rouge==1.1
pyarrow==1.0.1
pycparser==2.19
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pyspin==1.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2019.3
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2019.11.1
requests==2.22.0
s3transfer==0.2.1
sacremoses==0.0.35
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
sentencepiece==0.1.83
six==1.13.0
smart-open==1.9.0
soupsieve==1.9.5
spacy==2.2.3
srsly==0.2.0
terminado==0.8.3
testpath==0.4.4
textdistance==4.1.5
thinc==7.3.1
torch==1.3.1
torchvision==0.4.2
tornado==6.0.4
tqdm==4.38.0
traitlets==4.3.3
transformers==2.1.1
urllib3==1.25.7
urwid==2.1.0
wasabi==0.4.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
